Remove old banner when update banner

+++ api/app/Http/Controllers/AdminController.php
@@ -305,7 +305,9 @@ class AdminController extends Controller
             $file = $request->file('banner');
             $filename = 'event_' . $request->event_id . '_' . time() . '.' . $file->getClientOriginalExtension();
             $path = Storage::putFileAs('banners', $file, $filename);
-            Storage::delete($event->banner);
+            if ($event->banner && $event->banner != "banners/default.png") {
+                Storage::delete($event->banner);
+            }
             $event->banner = $path;
         }
